+++ changelog.md
@@ -22,6 +22,11 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 -
 
+## [2.6.0] - 2025-07-06
+
+### Fixed
+
+- fixed issue with markdown viewer not working on news and events
 ## [2.5.0] - 2025-07-06
 
 ### Added